+++ WinServiceBasic1/bin/Debug/WinServiceBasic1_bin_install.ps1
@@ -1,0 +1,3 @@
+cp WinServiceBasic1.exe c:\bin\
+cp NLog.config c:\bin\
+cp NLog.dll c:\bin\